- Restructure Folder Settings. - Implement Home Page Job Layout.
