+++ src/components/pages/Home.js
@@ -1,7 +1,7 @@
 import { useEffect } from "react"
-import SearchBox from "./SearchBox"
-import Jobs from "./Jobs"
-import Page from "./Page"
+import SearchBox from "../search/SearchBox"
+import Jobs from "../jobs/Jobs"
+import Page from "../pages/Page"
 
 const Home = () => {
   useEffect(() => {
