+++ src/components/pages/Terms.js
